[web] Add the Mapbox in the fullscreen background
