+++ web/package.json
@@ -5,6 +5,7 @@
   "scripts": {
     "dev": "webpack-dev-server --open --config webpack.dev.js",
     "build": "webpack --config webpack.prod.js",
+    "start": "npm run build && serve --single dist",
     "lint": "eslint src"
   },
   "eslintConfig": {
@@ -26,21 +27,27 @@
     "@turf/bbox": "~6.0.1",
     "@turf/helpers": "~6.1.4",
     "babel-eslint": "~10.0.1",
-    "babel-loader": "~8.0.5",
-    "clean-webpack-plugin": "~2.0.1",
+    "babel-loader": "~8.0.6",
+    "clean-webpack-plugin": "~2.0.2",
+    "css-loader": "~2.1.1",
     "dotenv-webpack": "~1.7.0",
     "file-loader": "~3.0.1",
     "html-webpack-plugin": "~3.2.0",
     "is-empty": "~1.2.0",
     "lodash.isequal": "~4.5.0",
+    "node-sass": "~4.12.0",
     "notistack": "~0.8.2",
     "places.js": "~1.16.4",
     "prop-types": "~15.7.2",
     "query-string": "~6.5.0",
     "react": "~16.8.6",
     "react-dom": "~16.8.6",
+    "react-jss": "~8.6.1",
     "react-map-gl": "~4.1.1",
+    "sass-loader": "~7.1.0",
+    "serve": "~11.0.0",
     "shortid": "~2.2.14",
+    "style-loader": "~0.23.1",
     "viewport-mercator-project": "~6.1.1",
     "webpack": "~4.31.0",
     "webpack-cli": "~3.3.2",
