+++ web/public/index.html
@@ -5,6 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/mapbox-gl@0.53.1/dist/mapbox-gl.min.css" integrity="sha384-p5EgjskuxekqL18U4iehyUHUMJNIH5rntB/wfABMWVHZPRIRQNYsLwAUdOQ6otH/" crossorigin="anonymous">
+    <!-- jss-insertion-point -->
     <title>Bistrotime</title>
   </head>
   <body>
