+++ web/src/components/bar.scss
@@ -1,0 +1,8 @@
+.Bar {
+  padding-top: 16px;
+  padding-left: 16px;
+  padding-right: 16px;
+  margin-bottom: 16px;
+  padding-bottom: 16px;
+  background-color: #f5f5f5;
+}
