+++ web/src/components/map.scss
@@ -1,0 +1,7 @@
+.MapContainer {
+  position: absolute;
+  top: 0px;
+  bottom: 0px;
+  width: 100%;
+  z-index: -10;
+}
