+++ web/src/components/pin.scss
@@ -1,0 +1,5 @@
+.Pin {
+  transform: translate(-10px, 20px);
+  cursor: pointer;
+  stroke: none;
+}
