+++ web/src/components/place.scss
@@ -1,0 +1,5 @@
+.algolia-places {
+  font-size: 14px;
+  margin-top: 8px;
+  font-family: "Roboto", "Helvetica", "Arial", sans-serif;
+}
