+++ web/src/index.js
@@ -1,15 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { SnackbarProvider } from 'notistack';
-
 
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 
-ReactDOM.render((
-  <SnackbarProvider maxSnack={3}>
-    <App />
-  </SnackbarProvider>
-), document.getElementById('app'));
+ReactDOM.render((<App />), document.getElementById('app'));
 
 serviceWorker.unregister();
