+++ web/webpack.common.js
@@ -22,6 +22,10 @@ module.exports = {
           },
         },
       },
+      {
+        test: /\.scss$/,
+        use: ['style-loader', 'css-loader', 'sass-loader'],
+      },
     ],
   },
   externals: {
